+++ kealth-core/src/main/kotlin/io/github/marioalvial/kealth/core/HealthComponent.kt
@@ -2,8 +2,12 @@ package io.github.marioalvial.kealth.core
 
 import io.github.marioalvial.kealth.core.HealthStatus.UNHEALTHY
 import io.github.marioalvial.kealth.extensions.measureTimeMillisAndReturn
-import kotlinx.coroutines.*
-import kotlinx.coroutines.Dispatchers.Default
+import kotlinx.coroutines.CoroutineExceptionHandler
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers.IO
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
 
@@ -11,48 +15,88 @@ import kotlin.coroutines.EmptyCoroutineContext
  * Interface that abstracts a health component.
  * @property name Name of health component
  * @property criticalLevel Critical component level
+ * @property componentContext Set the context that `health()` will run
  */
-interface HealthComponent {
+abstract class HealthComponent {
 
-    val name: String
-    val criticalLevel: String
-
-    /**
-     * Handle response of healthCheck() method
-     * @return HealthStatus
-     */
-    suspend fun health(): HealthInfo {
-        val context = context() + Dispatchers.IO
-
-        return measureTimeMillisAndReturn {
-            runCatching { withContext(context) { doHealthCheck() } }
-                .fold(
-                    onSuccess = { it },
-                    onFailure = {
-                        val job = GlobalScope.launch(context) { handleFailure(it) }
-                        job.invokeOnCompletion { runBlocking { job.join() } }
-                        UNHEALTHY
-                    }
-                )
-        }
-            .let { HealthInfo(it.first, criticalLevel, it.second) }
+    abstract val name: String
+    abstract val criticalLevel: String
+    open var componentContext: CoroutineContext = EmptyCoroutineContext
+    private val parameters: MutableMap<Any, Any> = mutableMapOf()
+    private val context by lazy {
+        CoroutineExceptionHandler { ctx, exception -> handleCoroutineException(ctx, exception) } + componentContext
     }
 
     /**
-     * If healthCheck() throws exception or return HealthStatus.UNHEALTHY executes logic to handle failure.
+     * Handle response of `healthCheck()` method
+     * @return HealthStatus
+     */
+    suspend fun health(): HealthInfo = measureTimeMillisAndReturn {
+        runCatching { withContext(componentContext) { doHealthCheck() } }
+            .fold(
+                onSuccess = {
+                    if (it == UNHEALTHY) handleUnhealthyStatus()
+
+                    it
+                },
+                onFailure = {
+                    handleThrowable(it)
+
+                    UNHEALTHY
+                }
+            )
+    }
+        .let { HealthInfo(it.first, criticalLevel, it.second) }
+
+    private fun handleThrowable(throwable: Throwable) {
+        val errorScope = CoroutineScope(IO + context)
+
+        errorScope
+            .launch { handleException(throwable) }
+            .invokeOnCompletion { errorScope.cancel() }
+    }
+
+    /**
+     * If doHealthCheck() throws exception executes `handleException()` logic.
      * @param throwable Throwable
      */
-    fun handleFailure(throwable: Throwable)
+    abstract fun handleException(throwable: Throwable)
+
+    /**
+     * If coroutine execution throws exception should execute `handleCoroutineException()` logic.
+     * @param coroutineContext CoroutineContext
+     * @param exception Throwable
+     */
+    abstract fun handleCoroutineException(coroutineContext: CoroutineContext, exception: Throwable)
+
+    /**
+     * If doHealthCheck() returns UNHEALTHY status executes handleUnhealthyStatus logic.
+     */
+    open fun handleUnhealthyStatus() = Unit
 
     /**
      * Execute the component's health check logic
      * @return HealthStatus
      */
-    fun doHealthCheck(): HealthStatus
+    protected abstract fun doHealthCheck(): HealthStatus
 
     /**
-     * Set shared context between threads
+     * Set shared componentContext between threads
      * @return CoroutineContext
      */
-    fun context(): CoroutineContext = EmptyCoroutineContext
+
+    /**
+     * Add parameter to parameters map
+     * @param key Any
+     * @param value Any
+     */
+    protected fun addParameter(key: Any, value: Any) {
+        parameters[key] = value
+    }
+
+    /**
+     * Get parameters map
+     * @return Map<Any, Any>
+     */
+    protected fun parameters(): Map<Any, Any> = parameters.toMap()
 }