+++ kealth-core/src/test/kotlin/io/github/marioalvial/kealth/testing/HealthComponentA.kt
@@ -3,8 +3,9 @@ package io.github.marioalvial.kealth.testing
 import io.github.marioalvial.kealth.core.CriticalLevel
 import io.github.marioalvial.kealth.core.HealthComponent
 import io.github.marioalvial.kealth.core.HealthStatus
+import kotlin.coroutines.CoroutineContext
 
-class HealthComponentA : HealthComponent {
+class HealthComponentA : HealthComponent() {
 
     override val name = "component A"
     override val criticalLevel = CriticalLevel.HIGH
@@ -19,11 +20,15 @@ class HealthComponentA : HealthComponent {
         return HealthStatus.HEALTHY
     }
 
-    override fun handleFailure(throwable: Throwable) {
-            println("Starting handleFailure of component $name in thread ${Thread.currentThread().name}")
+    override fun handleException(throwable: Throwable) {
+            println("Starting handleException of component $name in thread ${Thread.currentThread().name}")
 
             Thread.sleep(100)
 
             println("Finish handle failure of component $name - 100ms")
+    }
+
+    override fun handleCoroutineException(coroutineContext: CoroutineContext, exception: Throwable) {
+        println("Coroutine throws exception with context $coroutineContext and error ${exception.printStackTrace()}")
     }
 }