+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 2.0.0
+> Published 28 July 2019
+
+#### Improvements:
+
+* Add custom coroutine scope to run the `aggregate()`
+
+#### Features:
+
+* Added `handleUnhealthyStatus()` to executes logic when `doHealthCheck()` returns UNHEALTHY but does not throw exception
+* Added `parameters` variable to allow the component to add parameters to be used in any handle method
+
+#### Bugfix:
+
+* Change context logic to be set one time with a variable instead of method
+
 # 1.0.9.5
 > Published 9 July 2019